azure_iot_solution/modules/orchestrator/requirements.txt
smbus2
requests
luma.oled
RPi.GPIO
pyaudio
pocketsphinx
azure-cognitiveservices-speech
python-dotenv==1.0.0
